feat: Add AboutUs component for displaying information about the company
This commit adds the AboutUs component to the project. The component displays information about the company, including the number of enrolled merchants, certified courses, and rewards and gift cards. It also includes a section for users to apply to become a merchant. The component is responsive and can be viewed on any device.